+++ ChangeLog.md
@@ -12,7 +12,17 @@ and this project adheres to
 
 ### Changed
 * The `normalize` argument to `freud.density.RDF` is now `normalization_mode`.
+* The ``neighbors`` argument to ``env_neighbors`` for ``EnvironmentMotifMatch`` class.
+* The ``neighbors`` argument to ``cluster_neighbors`` for ``EnvironmentCluster`` class.
 * The `freud.order.Nematic` method no longer requires `u` ("nematic director") and uses orientation vectors instead of quaternions in the `compute` method.
+
+### Removed
+* The `global_search` flag in `freud.environment.EnvironmentCluster`.
+
+## v2.13.0 -- YYYY-MM-DD
+
+### Added
+* Filter neighborlists with `freud.locality.FilterSANN`.
 
 ## v2.12.1 -- 2022-12-05